# [dsd] optimize parsing logic for performance