### What does this PR do?

I continued profiling `pkg/dogstatsd` to reduce heap allocations and CPU usage. 

Most of the work is removing `bytes.Split` occurrences that allocated useless `[][]byte`s, and directly tokenkize the input `[]byte`.

The remaining allocations are individual `string`s and `[]string`s, that can't be avoided by design.

#### parseTags: -25% cpu, -80 B/op
```
BenchmarkCurrent-4   	 5000000	       315 ns/op	     160 B/op	       5 allocs/op
BenchmarkSplit2-4    	 5000000	       235 ns/op	      80 B/op	       4 allocs/op
```

#### parseMetricMessage: -38% CPU, - 128 B/op
```
BenchmarkMetricCurrent-4   	 2000000	       609 ns/op	     328 B/op	       8 allocs/op
BenchmarkMetricSplit2-4    	 5000000	       374 ns/op	     200 B/op	       5 allocs/op
```

Events parsing would also benefit from the same work, but that's for another day.